+++ src/components/common/index.js
@@ -4,3 +4,4 @@ export { default as Button } from "./button";
 export { default as Chip } from "./chip";
 export { default as Navigation } from "./navigation";
 export { default as Overlay } from "./overlay";
+export { default as Input } from "./input";
