+++ src/components/common/input/styles.js
@@ -1,0 +1,18 @@
+export const inputStyle = {
+  borderRadius: "11px",
+  padding: "12px",
+};
+
+export const iconStyle = {
+  paddingLeft: "16px",
+};
+
+export const containerStyle = {
+  borderColor: "#000",
+  color: "#000",
+};
+
+export const labelStyle = {
+  paddingBottom: "4px",
+  fontSize: "14px",
+};
